requirements.txt
aiogram==3.4.1
aiohttp==3.8.4
python-dotenv==1.0.1
